Initiative A: value/prioritization v1 (ADR 0039)
A1 (T-1007): priority becomes a live float64 [0,1] (was an inert int),
settable via gw ticket create/edit with range validation, default unset
(treated as 0), round-tripping through deterministic export/import. No
migration needed — SQLite's permissive INTEGER affinity stores REAL and
no rows carry a priority yet; scans now read NullFloat64.

A2 (T-1008): replace the scheduler's FIFO-by-id with orderByValue() —
the lexicographic priority path (per level: higher priority first, then
lower id for DFS/FIFO) over each node's root->node ancestor chain. Priority
is sibling-scoped by construction; the comparator is the single score()
seam the future multi-signal value model plugs into.

Tests: comparator cases (incl. the cross-root id tiebreak that
distinguishes per-level interleaving from a naive scheme) + an
order-by-value integration test.

+++ internal/exporter/exporter.go
@@ -24,7 +24,7 @@ type frontMatter struct {
 	Status         string   `yaml:"status"`
 	Assignee       *string  `yaml:"assignee"`
 	RequestedActor *string  `yaml:"requested_actor"`
-	Priority       *int     `yaml:"priority"`
+	Priority       *float64 `yaml:"priority"`
 	Labels         []string `yaml:"labels"`
 	Parent         *string  `yaml:"parent"`
 	DependsOn      []string `yaml:"depends_on"`
